storage, raftstore: move RegionInfoProvider into raftstore
