+++ tests/integrations/storage/test_seek_region.rs
@@ -5,8 +5,7 @@ use std::thread;
 use std::time::Duration;
 
 use test_raftstore::*;
-use tikv::raftstore::coprocessor::RegionInfoAccessor;
-use tikv::storage::kv::RegionInfoProvider;
+use tikv::raftstore::coprocessor::{RegionInfoAccessor, RegionInfoProvider};
 use tikv_util::collections::HashMap;
 use tikv_util::HandyRwLock;
